Move some tests out of test_pipeline.py and create MockDenseRetriever
