--- test/samples/pipeline/test_pipeline_faiss_indexing.yaml
@@ -1,30 +1,0 @@
-version: ignore
-
-components:
-  - name: DPRRetriever
-    type: DensePassageRetriever
-    params:
-      document_store: NewFAISSDocumentStore
-  - name: PDFConverter
-    type: PDFToTextConverter
-    params:
-      remove_numeric_tables: false
-  - name: Preprocessor
-    type: PreProcessor
-    params:
-      clean_whitespace: true
-  - name: NewFAISSDocumentStore
-    type: FAISSDocumentStore
-
-
-pipelines:
-  - name: indexing_pipeline
-    nodes:
-      - name: PDFConverter
-        inputs: [File]
-      - name: Preprocessor
-        inputs: [PDFConverter]
-      - name: DPRRetriever
-        inputs: [Preprocessor]
-      - name: NewFAISSDocumentStore
-        inputs: [DPRRetriever]
