--- test/samples/pipeline/test_pipeline_faiss_retrieval.yaml
@@ -1,18 +1,0 @@
-version: ignore
-
-components:
-  - name: DPRRetriever
-    type: DensePassageRetriever
-    params:
-      document_store: ExistingFAISSDocumentStore
-  - name: ExistingFAISSDocumentStore
-    type: FAISSDocumentStore
-    params:
-      faiss_index_path: 'existing_faiss_document_store'
-
-
-pipelines:
-  - name: query_pipeline
-    nodes:
-      - name: DPRRetriever
-        inputs: [Query]
